requirements.txt
Flask==3.0.3
selenium==4.23.1
python-dotenv==1.0.1
pandas==2.2.2
